tweaking embeds_many options

+++ lib/mongoid/associations/embeds_many.rb
@@ -156,10 +156,10 @@ module Mongoid #:nodoc:
       # Returns:
       #
       # The newly build target Document.
-      def nested_build(attributes, options = nil)
+      def nested_build(attributes, options = {})
         attributes.each do |index, attrs|
           if document = detect { |document| document._index == index.to_i }
-            if options[:allow_destroy] && attrs['_destroy']
+            if options && options[:allow_destroy] && attrs['_destroy']
               @target.delete(document)
               document.destroy
             else
